Update Fire Mage tests and gear configuration
- Added Fire Mage test results (TestFire.results)
- Renamed and updated fire_test.go
- Updated test expectations across all specs after test run
- Updated gear sets and APL configurations for mage specs
