--- sim/core/TestProtoVersioning.results
@@ -1,1 +1,0 @@
-saved_version_number: 1
