+++ sim/death_knight/unholy/TestUnholy.results
@@ -232,14 +232,6 @@ dps_results: {
  }
 }
 dps_results: {
- key: "TestUnholy-AllItems-Fen-Yu,FuryofXuen-102248"
- value: {
-  dps: 144104.78449
-  tps: 107429.31639
-  hps: 1977.00579
- }
-}
-dps_results: {
  key: "TestUnholy-AllItems-FleetPrimalDiamond"
  value: {
   dps: 137669.08128
@@ -261,14 +253,6 @@ dps_results: {
   dps: 146760.91993
   tps: 107650.60792
   hps: 1918.33957
- }
-}
-dps_results: {
- key: "TestUnholy-AllItems-Gong-Lu,StrengthofXuen-102249"
- value: {
-  dps: 154280.61312
-  tps: 114874.39936
-  hps: 1971.98108
  }
 }
 dps_results: {
@@ -400,22 +384,6 @@ dps_results: {
  }
 }
 dps_results: {
- key: "TestUnholy-AllItems-Qian-Le,CourageofNiuzao-102245"
- value: {
-  dps: 136805.75216
-  tps: 100848.98726
-  hps: 1948.87715
- }
-}
-dps_results: {
- key: "TestUnholy-AllItems-Qian-Ying,FortitudeofNiuzao-102250"
- value: {
-  dps: 143883.02392
-  tps: 105573.25569
-  hps: 1915.84087
- }
-}
-dps_results: {
  key: "TestUnholy-AllItems-Renataki'sSoulCharm-96741"
  value: {
   dps: 134358.13525
@@ -541,14 +509,6 @@ dps_results: {
   dps: 133809.21535
   tps: 98302.00168
   hps: 1854.53229
- }
-}
-dps_results: {
- key: "TestUnholy-AllItems-Xing-Ho,BreathofYu'lon-102246"
- value: {
-  dps: 138342.95297
-  tps: 101709.90404
-  hps: 1966.52707
  }
 }
 dps_results: {
