+++ sim/priest/shadow/TestShadow.results
@@ -689,13 +689,6 @@ dps_results: {
  }
 }
 dps_results: {
- key: "TestShadow-AllItems-Fen-Yu,FuryofXuen-102248"
- value: {
-  dps: 83554.88335
-  tps: 78780.70052
- }
-}
-dps_results: {
  key: "TestShadow-AllItems-FlashfrozenResinGlobule-100951"
  value: {
   dps: 88805.1455
@@ -777,13 +770,6 @@ dps_results: {
  value: {
   dps: 84354.31927
   tps: 79047.87448
- }
-}
-dps_results: {
- key: "TestShadow-AllItems-Gong-Lu,StrengthofXuen-102249"
- value: {
-  dps: 83554.88335
-  tps: 78780.70052
  }
 }
 dps_results: {
@@ -1620,20 +1606,6 @@ dps_results: {
  }
 }
 dps_results: {
- key: "TestShadow-AllItems-Qian-Le,CourageofNiuzao-102245"
- value: {
-  dps: 82264.59831
-  tps: 77466.66561
- }
-}
-dps_results: {
- key: "TestShadow-AllItems-Qian-Ying,FortitudeofNiuzao-102250"
- value: {
-  dps: 80949.1112
-  tps: 76352.94683
- }
-}
-dps_results: {
  key: "TestShadow-AllItems-Qin-xi'sPolarizingSeal-87075"
  value: {
   dps: 84796.75078
@@ -2296,13 +2268,6 @@ dps_results: {
  value: {
   dps: 114676.29499
   tps: 107140.75007
- }
-}
-dps_results: {
- key: "TestShadow-AllItems-Xing-Ho,BreathofYu'lon-102246"
- value: {
-  dps: 108033.28993
-  tps: 102699.85994
  }
 }
 dps_results: {
