+++ sim/warrior/fury/TestFury.results
@@ -192,13 +192,6 @@ dps_results: {
  }
 }
 dps_results: {
- key: "TestFury-AllItems-Fen-Yu,FuryofXuen-102248"
- value: {
-  dps: 147441.52381
-  tps: 94386.93656
- }
-}
-dps_results: {
  key: "TestFury-AllItems-FleetPrimalDiamond"
  value: {
   dps: 140280.01216
@@ -217,13 +210,6 @@ dps_results: {
  value: {
   dps: 149640.30832
   tps: 93253.9133
- }
-}
-dps_results: {
- key: "TestFury-AllItems-Gong-Lu,StrengthofXuen-102249"
- value: {
-  dps: 154225.71926
-  tps: 100213.12864
  }
 }
 dps_results: {
@@ -311,20 +297,6 @@ dps_results: {
  }
 }
 dps_results: {
- key: "TestFury-AllItems-Qian-Le,CourageofNiuzao-102245"
- value: {
-  dps: 140979.77638
-  tps: 88927.31392
- }
-}
-dps_results: {
- key: "TestFury-AllItems-Qian-Ying,FortitudeofNiuzao-102250"
- value: {
-  dps: 144065.693
-  tps: 90813.32715
- }
-}
-dps_results: {
  key: "TestFury-AllItems-Renataki'sSoulCharm-96741"
  value: {
   dps: 139949.09209
@@ -392,13 +364,6 @@ dps_results: {
  value: {
   dps: 140369.22251
   tps: 89184.07151
- }
-}
-dps_results: {
- key: "TestFury-AllItems-Xing-Ho,BreathofYu'lon-102246"
- value: {
-  dps: 141245.39134
-  tps: 89180.10613
  }
 }
 dps_results: {
