+++ ui/mage/arcane/gear_sets/p1_bis.gear.json
@@ -1,20 +1,20 @@
 {
-	"items": [
-		{ "id": 87008, "gems": [76885, 76643], "reforging": 144 },
-		{ "id": 87076, "reforging": 147 },
-		{ "id": 87011, "enchant": 4806, "gems": [76643] },
-		{ "id": 90512, "enchant": 4892, "reforging": 153, "tinker": 4897 },
-		{ "id": 87169, "enchant": 4419, "gems": [76672, 76700] },
-		{ "id": 86948, "enchant": 4414, "gems": [0], "reforging": 153 },
-		{ "id": 87007, "enchant": 4433, "gems": [0], "reforging": 140, "tinker": 4898 },
-		{ "id": 86981, "gems": [76700, 76643, 76700], "reforging": 154, "tinker": 4223 },
-		{ "id": 87009, "enchant": 4825, "gems": [76672], "reforging": 153 },
-		{ "id": 87162, "enchant": 4426, "gems": [76700], "reforging": 151 },
-		{ "id": 86949, "reforging": 147 },
-		{ "id": 90511, "reforging": 119 },
-		{ "id": 87065, "reforging": 154 },
-		{ "id": 87175 },
-		{ "id": 90513, "enchant": 4442, "gems": [76643], "reforging": 145 },
-		{ "id": 86960, "enchant": 4434, "reforging": 147 }
-	]
-}
+ "items": [
+        {"id":87008,"gems":[76885,76643],"reforging":144},
+        {"id":87076,"reforging":147},
+        {"id":87011,"enchant":4806,"gems":[76643]},
+        {"id":90512,"enchant":4892,"reforging":153,"tinker":4897},
+        {"id":87169,"enchant":4419,"gems":[76672,76700]},
+        {"id":86948,"enchant":4414,"gems":[0],"reforging":153},
+        {"id":87007,"enchant":4433,"gems":[0],"reforging":140,"tinker":4898},
+        {"id":86981,"gems":[76700,76643,76700],"reforging":154,"tinker":4223},
+        {"id":87009,"enchant":4825,"gems":[76672],"reforging":153},
+        {"id":87162,"enchant":4426,"gems":[76700],"reforging":151},
+        {"id":86949,"reforging":147},
+        {"id":90511,"reforging":119},
+        {"id":87065,"reforging":154},
+        {"id":87175},
+        {"id":90513,"enchant":4442,"gems":[76643],"reforging":145},
+        {"id":86960,"enchant":4434,"reforging":147}
+      ]
+    }
